fixed for two way communication between main and mem, just have to perform actual functionality now

+++ mem.c
@@ -13,24 +13,28 @@ int main(int argc, char **argv){
     char *pipe = "rdpipe";
     char buf[1024];
 
-    fd = open(pipe, O_RDONLY);
-    read(fd, buf, 1024);
+    if((fd = open(pipe, O_RDONLY)) > 0){
+      read(fd, buf, 1024);
  
-    sleep(5);
+      sleep(5);
 
-    printf("Received message: %s\n", buf);
-    close(fd);
+      printf("Received message: %s\n", buf);
+      close(fd);
 
-    /*char *pipe2 = "wrpipe";
+      unlink(pipe);
 
-    mkfifo(pipe2, 0666);
+      char *pipe2 = "wrpipe";
 
-    fd = open(pipe2, O_WRONLY);
-    write(fd, "bye", sizeof("bye"));
-    close(fd);
+      mkfifo(pipe2, 0666);
+
+      fd = open(pipe2, O_WRONLY);
+      write(fd, "bye", sizeof("bye"));
+      close(fd);
     
-    unlink(pipe2);
-*/
+      unlink(pipe2);
+    }else{
+      continue;
+    }
   }
 
   exit(0);
